+++ types/node/events.d.ts
@@ -1,8 +1,16 @@
 declare module "events" {
     class internal extends NodeJS.EventEmitter { }
 
+    interface NodeEventTarget {
+        once(event: string | symbol, listener: (...args: any[]) => void): this;
+    }
+
+    interface DomEventTarget {
+        addEventListener(event: string, listener: (...args: any[]) => void, opts: { once: boolean }): any;
+    }
+
     namespace internal {
-        function once(emitter: EventEmitter, event: string | symbol): Promise<any[]>;
+        function once(emitter: NodeEventTarget | DomEventTarget, event: string | symbol): Promise<any[]>;
          class EventEmitter extends internal {
             /** @deprecated since v4.0.0 */
             static listenerCount(emitter: EventEmitter, event: string | symbol): number;
