+++ types/node/globals.d.ts
@@ -341,7 +341,7 @@ declare class Buffer extends Uint8Array {
      * @param encoding encoding used to evaluate (defaults to 'utf8')
      */
     static byteLength(
-        string: string | NodeJS.TypedArray | DataView | ArrayBuffer | SharedArrayBuffer,
+        string: string | NodeJS.ArrayBufferView | ArrayBuffer | SharedArrayBuffer,
         encoding?: BufferEncoding
     ): number;
     /**
@@ -651,9 +651,7 @@ declare namespace NodeJS {
 
     interface ReadWriteStream extends ReadableStream, WritableStream { }
 
-    interface Events extends EventEmitter { }
-
-    interface Domain extends Events {
+    interface Domain extends EventEmitter {
         run<T>(fn: (...args: any[]) => T, ...args: any[]): T;
         add(emitter: EventEmitter | Timer): void;
         remove(emitter: EventEmitter | Timer): void;
@@ -1152,4 +1150,5 @@ declare namespace NodeJS {
     }
 
     type TypedArray = Uint8Array | Uint8ClampedArray | Uint16Array | Uint32Array | Int8Array | Int16Array | Int32Array | Float32Array | Float64Array;
+    type ArrayBufferView = TypedArray | DataView;
 }
