+++ types/node/ts3.2/index.d.ts
@@ -16,5 +16,6 @@
 /// <reference path="../base.d.ts" />
 
 // TypeScript 3.2-specific augmentations:
+/// <reference path="fs.d.ts" />
 /// <reference path="util.d.ts" />
 /// <reference path="globals.d.ts" />
