+++ types/node/ts3.2/node-tests.ts
@@ -25,6 +25,7 @@ import '../worker_threads';
 import '../zlib';
 
 import { types } from 'util';
+import { BigIntStats, statSync, Stats } from 'fs';
 
 //////////////////////////////////////////////////////////
 /// Global Tests : https://nodejs.org/api/global.html  ///
@@ -48,6 +49,12 @@ import { types } from 'util';
         // $ExpectType number
         const b = value;
     }
+}
+
+// FS Tests
+{
+    const bigStats: BigIntStats = statSync('.', { bigint: true });
+    const anyStats: Stats | BigIntStats = statSync('.', { bigint: Math.random() > 0.5 });
 }
 
 // Global Tests
